// src/app/models/api.models.ts
export interface Order {
  id: string;
  customerId: string;
  customerName?: string;
  customerEmail?: string;
  sellerId: string;
  status: 'PENDING' | 'PROCESSING' | 'SHIPPED' | 'DELIVERED' | 'CANCELLED';
  type: 'PURCHASE' | 'RENTAL';
  items: OrderItem[];
  subtotal: number;
  tax: number;
  shippingCost: number;
  totalAmount: number;
  shippingAddress: string;
  trackingNumber?: string;
  estimatedDelivery?: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItem {
  productId: string;
  productName: string;
  quantity: number;
  price: number;
  type: 'PURCHASE' | 'RENTAL';
  rentalDays?: number;
}
// ...
export interface Wallet {
  id: string;
  userId: string;
  balance: number;
  loyaltyPoints: number;
  updatedAt: string;
}

export interface WalletTransaction {
  id: string;
  walletId: string;
  type: 'CREDIT' | 'DEBIT';
  amount: number;
  description: string;
  status: 'COMPLETED' | 'PENDING' | 'FAILED';
  createdAt: string;
}

export interface AddFundsDto {
  amount: number;
  source: 'CARD' | 'BANK_TRANSFER';
}

// ========== CART ==========
export interface CartItem {
  productId: string;
  productName: string;
  price: number;
  quantity: number;
  image: string;
  type: 'PURCHASE' | 'RENTAL';
  rentalDays?: number;
}